local -> def and require -> autoload
I've also moved some require calls to the module macro to keep them
together. By using autoload (a feature of Aniseed) you get JIT requiring
of modules, so if a feature is only used rarely but requires a large
module to be loaded it'll now be deferred until use.

This works in a similar way to Vim Script's autoload files.